fixed directory structure again
