+++ .pathplanner/settings.json
@@ -1,5 +1,5 @@
 {
-  "robotWidth": 0.75,
+  "robotWidth": 1.0,
   "robotLength": 1.0,
   "holonomicMode": true,
   "generateJSON": false,
