--- SwerveDrive-old-import/.pathplanner/settings.json
@@ -1,7 +1,0 @@
-{
-  "robotWidth": 1.0,
-  "robotLength": 1.0,
-  "holonomicMode": true,
-  "generateJSON": false,
-  "generateCSV": false
-}
