+++ src/main/deploy/example.txt
